requirements.txt
apache-airflow-providers-mongo==3.0.0
python-docx
PyPDF2
linkedin-api
pydantic
python-jobspy
openai
pinecone-client
